requirements.txt
python-telegram-bot>=20.0
APScheduler>=3.10.0
requests>=2.31.0
python-dotenv>=1.0.0
xmltodict>=0.13.0
pytest>=7.4.0
pytz>=2023.3
